requirements.txt
alabaster==0.7.12
arrow==1.2.3
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
b2sdk==1.19.0
Babel==2.10.3
black==22.10.0
boto3==1.24.89
botocore==1.27.89
build==0.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
codacy-coverage==1.3.11
coverage==6.5.0
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.2
django-backblaze-b2==3.1.0
django-bootstrap5==22.1
django-cleanup==6.0.0
django-environ==0.9.0
django-filter==22.1
django-storages==1.13.1
django-tables2==2.4.1
docutils==0.19
exceptiongroup==1.0.4
filelock==3.8.0
flake8==5.0.4
gunicorn==20.1.0
h11==0.14.0
idna==3.4
imagesize==1.4.1
importlib-resources==5.10.0
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
jmespath==1.0.1
logfury==1.0.1
MarkupSafe==2.1.1
mccabe==0.7.0
mock==4.0.3
mypy-extensions==0.4.3
outcome==1.2.0
packaging==21.3
pathspec==0.10.1
pep517==0.13.0
Pillow==9.4.0
pip-tools==6.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.4
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
selenium==4.6.0
sentry-sdk==1.9.10
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
tomli==2.0.1
tox==3.26.0
tox-gh-actions==2.10.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.5
webdriver-manager==3.8.5
whitenoise==6.2.0
wsproto==1.2.0
yamllint==1.28.0
